fix(cli-sub-agent): extend SA model-pin inheritance to review and debate (#1741)
Round-1 review found `csa run` honored the inherited CSA_MODEL_SPEC subtree
pin via resolve_handle_run_model_pin, but `csa review` and `csa debate`
ignored it: build_and_validate_executor resolved tool/model from
args/tier/defaults with no inherited-pin lookup, contradicting the
subtree_model_pin_prompt_guard advertised to agents. A pinned Layer-1
agent running `csa review --sa-mode true` (no --model-spec) would silently
lose the pin and re-route to the tier's first tool (dead gemini-cli in
tier-4-critical) -> the exact #1741 hang.

- review_cmd_subtree_pin.rs: apply_inherited_model_pin in the review
  executor build path (review_cmd_execute.rs:216/236).
- debate_cmd_subtree_pin.rs: same for the debate executor (debate_cmd.rs:325).
- debate_cmd_question.rs: extract question handling from the debate_cmd.rs
  monolith to keep modules under the token-budget gate after the new code.
- run_cmd_model_pin_tests.rs: move the pin unit tests out of
  run_cmd_model_pin.rs (same gate reason).

An explicit child `--model-spec` still overrides; unpinned dispatch keeps
tier auto-routing. Addresses the sole round-1 review finding on #1741.
